add lowercaser and tests
