+++ tokenizers/src/lib.rs
@@ -26,8 +26,8 @@ use cjk::ChineseTokenizer;
 use code::CodeTokenizer;
 use lindera::{LinderaChineseTokenizer, LinderaJapaneseTokenizer, LinderaKoreanTokenizer};
 use tantivy::tokenizer::{
-    AsciiFoldingFilter, LowerCaser, NgramTokenizer, RawTokenizer, RemoveLongFilter,
-    SimpleTokenizer, Stemmer, TextAnalyzer, TokenizerManager,
+    AsciiFoldingFilter, Language, LowerCaser, NgramTokenizer, RawTokenizer, RemoveLongFilter,
+    SimpleTokenizer, Stemmer, TextAnalyzer, TokenizerManager, WhitespaceTokenizer,
 };
 use tracing::info;
 
@@ -43,9 +43,27 @@ pub fn create_tokenizer_manager(search_tokenizers: Vec<&SearchTokenizer>) -> Tok
 
     for search_tokenizer in search_tokenizers {
         let tokenizer_option = match search_tokenizer {
+            SearchTokenizer::Default => Some(
+                TextAnalyzer::builder(SimpleTokenizer::default())
+                    .filter(RemoveLongFilter::limit(DEFAULT_REMOVE_TOKEN_LENGTH))
+                    .filter(LowerCaser)
+                    .build(),
+            ),
             SearchTokenizer::Raw => Some(
                 TextAnalyzer::builder(RawTokenizer::default())
                     .filter(RemoveLongFilter::limit(DEFAULT_REMOVE_TOKEN_LENGTH))
+                    .build(),
+            ),
+            SearchTokenizer::Lowercase => Some(
+                TextAnalyzer::builder(RawTokenizer::default())
+                    .filter(RemoveLongFilter::limit(DEFAULT_REMOVE_TOKEN_LENGTH))
+                    .filter(LowerCaser)
+                    .build(),
+            ),
+            SearchTokenizer::WhiteSpace => Some(
+                TextAnalyzer::builder(WhitespaceTokenizer::default())
+                    .filter(RemoveLongFilter::limit(DEFAULT_REMOVE_TOKEN_LENGTH))
+                    .filter(LowerCaser)
                     .build(),
             ),
             SearchTokenizer::ChineseCompatible => Some(
@@ -91,6 +109,13 @@ pub fn create_tokenizer_manager(search_tokenizers: Vec<&SearchTokenizer>) -> Tok
                     .filter(LowerCaser)
                     .build(),
             ),
+            SearchTokenizer::EnStem => Some(
+                TextAnalyzer::builder(SimpleTokenizer::default())
+                    .filter(RemoveLongFilter::limit(DEFAULT_REMOVE_TOKEN_LENGTH))
+                    .filter(LowerCaser)
+                    .filter(Stemmer::new(Language::English))
+                    .build(),
+            ),
             SearchTokenizer::Stem { language } => Some(
                 TextAnalyzer::builder(SimpleTokenizer::default())
                     .filter(RemoveLongFilter::limit(DEFAULT_REMOVE_TOKEN_LENGTH))
@@ -105,7 +130,6 @@ pub fn create_tokenizer_manager(search_tokenizers: Vec<&SearchTokenizer>) -> Tok
                     .filter(LowerCaser)
                     .build(),
             ),
-            _ => None,
         };
 
         if let Some(text_analyzer) = tokenizer_option {
